Fix size types
Upstream pull request:
https://github.com/heimdal/heimdal/pull/354

(parts of cherry picked Samba commit 72979d1d60ca2eab1e7903c2e77b8cca69667691)

+++ lib/asn1/gen.c
@@ -539,7 +539,7 @@ is_primitive_type(int type)
 }
 
 static void
-space(int level)
+space(size_t level)
 {
     while(level-- > 0)
 	fprintf(headerfile, "  ");
@@ -568,7 +568,7 @@ have_ellipsis(Type *t)
 }
 
 static void
-define_asn1 (int level, Type *t)
+define_asn1 (size_t level, Type *t)
 {
     switch (t->type) {
     case TType:
